# Run migrations once per package 