--- a/internal/database/dbtest/dbtest.go
+++ b/internal/database/dbtest/dbtest.go
@@ -2,10 +2,12 @@ package dbtest
 
 import (
 	"database/sql"
+	"hash/fnv"
 	"math/rand"
 	"net/url"
 	"os"
 	"strconv"
+	"sync"
 	"testing"
 	"time"
 
@@ -56,27 +58,16 @@ func NewDB(t testing.TB, dsn string) *sql.DB {
 		}
 	}
 
+	initTemplateDB(t, config)
+
 	rng := rand.New(rand.NewSource(time.Now().UnixNano()))
 	dbname := "sourcegraph-test-" + strconv.FormatUint(rng.Uint64(), 10)
 
 	db := dbConn(t, config)
-	dbExec(t, db, `CREATE DATABASE `+pq.QuoteIdentifier(dbname))
+	dbExec(t, db, `CREATE DATABASE `+pq.QuoteIdentifier(dbname)+` TEMPLATE `+pq.QuoteIdentifier(templateDBName()))
 
 	config.Path = "/" + dbname
 	testDB := dbConn(t, config)
-
-	for _, database := range []*dbconn.Database{
-		dbconn.Frontend,
-		dbconn.CodeIntel,
-	} {
-		m, err := dbconn.NewMigrate(testDB, database)
-		if err != nil {
-			t.Fatalf("failed to construct migrations: %s", err)
-		}
-		if err = dbconn.DoMigrate(m); err != nil {
-			t.Fatalf("failed to apply migrations: %s", err)
-		}
-	}
 
 	t.Cleanup(func() {
 		defer db.Close()
@@ -94,6 +85,56 @@ func NewDB(t testing.TB, dsn string) *sql.DB {
 	})
 
 	return testDB
+}
+
+var templateOnce sync.Once
+
+// initTemplateDB creates a template database with a fully migrated schema for the
+// current package. New databases can then do a cheap copy of the migrated schema
+// rather than running the full migration every time.
+func initTemplateDB(t testing.TB, config *url.URL) {
+	templateOnce.Do(func() {
+		templateName := templateDBName()
+		db := dbConn(t, config)
+		dbExec(t, db, `DROP DATABASE IF EXISTS `+pq.QuoteIdentifier(templateName))
+		dbExec(t, db, `CREATE DATABASE `+pq.QuoteIdentifier(templateName))
+		defer db.Close()
+
+		cfgCopy := *config
+		cfgCopy.Path = "/" + templateName
+		templateDB := dbConn(t, &cfgCopy)
+		defer templateDB.Close()
+
+		for _, database := range []*dbconn.Database{
+			dbconn.Frontend,
+			dbconn.CodeIntel,
+		} {
+			m, err := dbconn.NewMigrate(templateDB, database)
+			if err != nil {
+				t.Fatalf("failed to construct migrations: %s", err)
+			}
+			if err = dbconn.DoMigrate(m); err != nil {
+				t.Fatalf("failed to apply migrations: %s", err)
+			}
+		}
+		dbExec(t, db, killClientConnsQuery, templateName)
+	})
+}
+
+// templateDBName returns the name of the template database
+// for the currently running package.
+func templateDBName() string {
+	return "sourcegraph-test-template-" + wdHash()
+}
+
+// wdHash returns a hash of the current working directory.
+// This is useful to get a stable identifier for the package running
+// the tests.
+func wdHash() string {
+	h := fnv.New64()
+	wd, _ := os.Getwd()
+	h.Write([]byte(wd))
+	return strconv.Itoa(int(h.Sum64()))
 }
 
 func dbConn(t testing.TB, cfg *url.URL) *sql.DB {