Previously, every time `NewDB()` was called, we would create a new
database and run the migration. We can significantly reduce the work
done by `NewDB()` by running the schema migration once, then creating new
test DBs using the first DB as a template. Creating a new DB from a
template takes ~500ms on my machine vs 6.5 seconds when running the
migrations. 

Note that this implementation creates a new template database per
package. Since go runs tests for each package as a separate process and
in parallel, it would be difficult to synchronise creating the template
DB across packages. However, I'd love suggestions to make this better.

This change will continue to improve performance as we migrate from
`dbtesting` to `dbtest`. And, since this makes creating DBs cheaper,
it encourages more parallelization. 

Before:
```
❯ go test ./internal/database
ok      github.com/sourcegraph/sourcegraph/internal/database    110.768s
```

After:
```
❯ go test ./internal/database
ok      github.com/sourcegraph/sourcegraph/internal/database    82.559s
```


<!-- Reminder: Have you updated the changelog and relevant docs (user docs, architecture diagram, etc) ? -->