@camdencheek 2021-05-25
@mrnugget Anecdotally, converting [internal/database/search_contexts_test.go](https://sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/internal/database/search_contexts_test.go?subtree=true) to use `NewDB()` and `t.Parallel()` rather than `dbtesting.SetupGlobalTestDB()` cut runtime for those tests from 14s to 4s (9 tests total, each calling `dbtesting.SetupGlobalTestDB()`. Now, if you're only calling `dbtesting.SetupGlobalTestDB()` once _per package_, it is likely still faster than running `NewDB()` once _per test_, but YMMV, especially if you also parallelize tests that couldn't be before