@camdencheek 2021-05-25
In addition to this, I plan to spend a little time looking at maintaining a pool of migrated DBs that we can just truncate and use. Not sure if performance will actually be better than the 0.5s template clone, but it seems worth taking a look at if we can make creating a test db negligibly cheap